Conversation opened. 1 read message.
Skip to content
Using Gmail with screen readers
18 of 35,436
android - Hiding views in RecyclerView - Stack Overflow
Inbox

Dan Misra <danmisra0@gmail.com>
Sat, Feb 22, 9:15 AM (1 day ago)
to me

https://stackoverflow.com/questions/27574805/hiding-views-in-recyclerview/35480247#35480247

* How to check bookmarks for a script line the user clicks on.

We can alternate between GONE / hidden and VISIBLE recycler view items.

The position number of the recycler view item prior to the one the user clicked on is 1 less than what the user clicked on.

In the hidden items we put the playcode, act_nr, scene_nr and play_line_nr of the visible recycler view item.

When the user clicks on a recycler view with position N item we find the recycler view item with position N-1 which is GONE/hidden. We read off the location details of the row.

Then we look up the bookmark table to check whether there is a bookmark. If there is then we show it to the user and allow the user to edit or remove it.

** How to show bookmark indicator.on script lines which have a bookmark.

Two options:

Option 1

Add.a bookmark column (bkmk_count) to the script table with default value 0. When a bookmark is added for a particular script line, increment the bookmark count. When a bookmark is removed for a particular line, decrement the counter.

Option 2

When generating the list which shows the script, left outer join to the bookmark table with a COUNT(bookmark_row_id) as bkmk_count where active_0_or_1 = 1, on play code, act_nr, scene_nr and play_line_nr. Obtain the bkmk_count for display when generating the script_line_list whixh feeds the recycler view as in option 1.

The bookmark count.for a line can be shown with a.number in square brackets, or with asterisks where the number of asterisks is equal to the number of bookmarks.

Preference

I prefer Option 2 because I don't want to add bookmark-related information.to the script table, and would prefer to keep them in separate tables which can be left outer join'ed readily, as explained above.

-end-