Removed the unnecessary extra visual spacing between lines by making the height of reference rows zero.
Included error catching for where there are two or more play lines between reference lines to use the closest previous reference line for bookmarks.

Added label to show current state of sound on/off toggle in the Home screen and in the Settings Home screen.

Added more cases of interrupting the text to speech when the user hits a button to move to a different screen to stop text to speech from prior screen when the user navigates to another screen.

Signed-off-by: Dan A <danabramsonwriter@gmail.com>