Now the script is returned as a 2-dimensional array list so that we have the name of the character as well as the script text.
Ultimately this will be used for the recycler view rather than the 1-dimensional version which will then be deprecated.

Next, the user will be given the opportunity to select which character they are playing. The user will then have the option of the play being read to them for characters other than their own, and a pause while the user reads their own line.

Next, we will want to be able to record the user's own voice for their own lines, and then analyse and give feedback on their delivery of their own lines.