Ensure the prologue check refers to the latest version of the play online. Also updated the local copy of the play. Noted that a number of features use the local copy of the play, while others features such as getting the script, the bookmarks, and the prologue check use the online version.
The local version is used for the user's current navigation position, the number of scenes in the act, the number of acts in the play, and other such information relating to play navigation. So we need to continue to maintain the local database and ideally synch is from the online version in the future, if possible.